+++ src/main/java/com/tsp/new_tsp_admin/api/domain/faq/AdminFaqEntity.java
@@ -1,11 +1,16 @@
 package com.tsp.new_tsp_admin.api.domain.faq;
 
 import com.tsp.new_tsp_admin.api.domain.common.NewCommonMappedClass;
+import com.tsp.new_tsp_admin.api.domain.model.AdminModelDTO;
+import com.tsp.new_tsp_admin.api.domain.model.AdminModelEntity;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
 
 import javax.persistence.*;
 import javax.validation.constraints.NotEmpty;
+
+import java.util.ArrayList;
+import java.util.List;
 
 import static javax.persistence.GenerationType.IDENTITY;
 
@@ -19,7 +24,7 @@ import static javax.persistence.GenerationType.IDENTITY;
 @Table(name = "tsp_notice")
 public class AdminFaqEntity extends NewCommonMappedClass {
     @Transient
-    private Integer rnum;
+    private Integer rowNum;
 
     @Id
     @GeneratedValue(strategy = IDENTITY)
@@ -41,4 +46,24 @@ public class AdminFaqEntity extends NewCommonMappedClass {
     @Column(name = "visible")
     @NotEmpty(message = "FAQ 노출 여부 선택은 필수입니다.")
     private String visible;
+
+    public static AdminFaqDTO toDto(AdminFaqEntity entity) {
+        return AdminFaqDTO.builder().idx(entity.getIdx())
+                .rowNum(entity.getRowNum())
+                .title(entity.getTitle())
+                .description(entity.getDescription())
+                .viewCount(entity.getViewCount())
+                .visible(entity.getVisible())
+                .creator(entity.getCreator())
+                .createTime(entity.getCreateTime())
+                .updater(entity.getUpdater())
+                .updateTime(entity.getUpdateTime())
+                .build();
+    }
+
+    public List<AdminFaqDTO> toDtoList(List<AdminFaqEntity> entityList) {
+        List<AdminFaqDTO> list = new ArrayList<>(entityList.size());
+        entityList.forEach(adminFaqEntity -> list.add(toDto(adminFaqEntity)));
+        return list;
+    }
 }
